[fei5850.3.creator] Get types to export
